Version 2.8 beta. Closes https://github.com/MizzleDK/MizuuApp/issues/27 and closes https://github.com/MizzleDK/MizuuApp/issues/18.

+++ src/com/miz/filesources/FileMovie.java
@@ -37,9 +37,11 @@ public class FileMovie extends MovieFileSource<File> {
 		File temp;
 		int count = dbMovies.size();
 		for (int i = 0; i < count; i++) {
-			temp = new File(dbMovies.get(i).getFilepath());
-			if (temp.exists() && dbMovies.get(i).isUnidentified())
-				db.deleteMovie(dbMovies.get(i).getRowId());
+			if (!dbMovies.get(i).isNetworkFile() && !dbMovies.get(i).isUpnpFile()) {
+				temp = new File(dbMovies.get(i).getFilepath());
+				if (temp.exists() && dbMovies.get(i).isUnidentified())
+					db.deleteMovie(dbMovies.get(i).getRowId());
+			}
 		}
 	}
